Rebase and bump version
In this patch I've bumped the version number as a minor update.
It was unclear whether this was a major or minor update because while
it changes the behavior of the topology state when server configs change
it doesn't do anything but prevent crashing the app.

For applications which depend on the crashing behavior (hopefully there
are none) then this is a major change.