Updates to support MongoDB 5.1+
In MongoDB 3.6, the `OP_MSG` and `OP_COMPRESSED` opcodes were added to
the wired message protocol. `OP_MSG` was added to standardize the
MongoDB message format, and `OP_COMPRESSED` takes things a step
further by compressing the message to increase network efficiency. As
of MongoDB 5.1, the old opcodes were removed, and all messages are
sent with either `OP_MSG` or `OP_COMPRESSED`.

Because this driver did not have the ability to send messages with the
`OP_MSG` opcode, it couldn't be used with any versions of MongoDB
greater than 5.1.

To bring this driver into the modern era, this PR ports over the
changes made in the [emqx fork](https://github.com/emqx/mongodb-erlang/)
of this driver to support the `OP_MSG` opcode.

Additionally, there are some slightly unrelated, but relevant changes
in this PR as well such as fixing some dialyzer errors and updating the
GitHub actions workflow definitions to get the tests up and running again.